+++ sdk/appconfiguration/Azure.ResourceManager.AppConfiguration/tests/Tests/PrivateEndpointConnectionOperationTests.cs
@@ -82,7 +82,7 @@ namespace Azure.ResourceManager.AppConfiguration.Tests
                             RequestMessage = "Please approve my connection",
                         }
                         },
-                    Subnet = new SubnetData() { Id = subnetID}
+                    Subnet = new SubnetData() { Id = subnetID }
                 };
                 PrivateEndpointResource = (await ResGroup.GetPrivateEndpoints().CreateOrUpdateAsync(WaitUntil.Completed, EndpointName, privateEndpointData)).Value;
                 List<AppConfigurationPrivateEndpointConnectionResource> connections = await ConfigStore.GetAppConfigurationPrivateEndpointConnections().GetAllAsync().ToEnumerableAsync();
@@ -90,6 +90,7 @@ namespace Azure.ResourceManager.AppConfiguration.Tests
             }
         }
 
+        [Ignore("Depend on Network which will block the pipeline to release new Network package, disable this case temporary")]
         [Test]
         public async Task DeleteTest()
         {
@@ -99,6 +100,7 @@ namespace Azure.ResourceManager.AppConfiguration.Tests
             Assert.AreEqual(404, exception.Status);
         }
 
+        [Ignore("Depend on Network which will block the pipeline to release new Network package, disable this case temporary")]
         [Test]
         public async Task GetTest()
         {
@@ -110,7 +112,7 @@ namespace Azure.ResourceManager.AppConfiguration.Tests
         [Test]
         public async Task GetAvailableLocationsTest()
         {
-            IEnumerable<AzureLocation> locations =  (await Connection.GetAvailableLocationsAsync()).Value;
+            IEnumerable<AzureLocation> locations = (await Connection.GetAvailableLocationsAsync()).Value;
 
             Assert.IsNotEmpty(locations);
         }